White wine quality dataset of the Portuguese "Vinho Verde" wine

A white wine tasting preference data used in the study of Cortez, Cerdeira, Almeida, Matos, and Reis 2009. This white wine contains 4898 white vinho verde wine samples and 12 variables including the tasting preference score of white wine and its physicochemical characteristics.

Usage

data(WhiteWine)

Format

A data frame with 4898 rows, quality score, and 11 variables of physicochemical properties of wines.

quality Tasting preference is a rating score provided by a minimum of three sensory with ordinal values from 0 (very bad) to 10 (excellent). The final sensory score is the median of these evaluations.

fixed.acidity The fixed acidity is the physicochemical property in unit (g(tartaric acid)/dm^3).

volatile.acidity The volatile acidity is in unit g(acetic acid)/dm^3.

citric.acid The citric acidity is in unit g/dm^3.

residual.sugar The residual sugar is in unit g/dm^3.

chlorides The chlorides is in unit g(sodium chloride)/dm^3.

free.sulfur.dioxide The free sulfur dioxide is in unit mg/dm^3.

total.sulfur.dioxide The total sulfur dioxide is in unit mg/dm^3.

density The density is in unit g/cm^3.

pH The wine's pH value.

sulphates The sulphates is in unit g(potassium sulphates)/dm^3.

alcohol The alcohol is in unit % vol.

References

Cortez, P., Cerdeira, A., Almeida, F., Matos, T., and Reis, J. (2009), “Modeling wine preferences by data mining from physicochemical properties,” Decision Support Systems, 47, 547–553. doi:10.1016/j.dss.2009.05.016

Examples

head(WhiteWine)
#>   quality fixed.acidity volatile.acidity citric.acid residual.sugar chlorides
#> 1       6           7.0             0.27        0.36           20.7     0.045
#> 2       6           6.3             0.30        0.34            1.6     0.049
#> 3       6           8.1             0.28        0.40            6.9     0.050
#> 4       6           7.2             0.23        0.32            8.5     0.058
#> 5       6           7.2             0.23        0.32            8.5     0.058
#> 6       6           8.1             0.28        0.40            6.9     0.050
#>   free.sulfur.dioxide total.sulfur.dioxide density   pH sulphates alcohol
#> 1                  45                  170  1.0010 3.00      0.45     8.8
#> 2                  14                  132  0.9940 3.30      0.49     9.5
#> 3                  30                   97  0.9951 3.26      0.44    10.1
#> 4                  47                  186  0.9956 3.19      0.40     9.9
#> 5                  47                  186  0.9956 3.19      0.40     9.9
#> 6                  30                   97  0.9951 3.26      0.44    10.1